feat: 테마 결과페이지 - 테마박스 컴포넌트 삭제
